Add catalog support
