+++ broker/lib/fabrik/BasePlatformManager.js
@@ -1,12 +1,22 @@
 'use strict';
 
+const _ = require('lodash');
+
 class BasePlatformManager {
   constructor(platform) {
     this.platform = platform;
   }
 
   getCatalog(catalog) {
-    return catalog;
+    const modifiedCatalog = _.cloneDeep(catalog);
+    const platform = this.platform;
+    _.remove(modifiedCatalog.services, function (service) {
+      _.remove(service.plans, function (plan) {
+        return !_.includes(_.get(plan, 'metadata.supportedPlatforms', ['cf']), platform);
+      });
+      return !_.includes(_.get(service, 'metadata.supportedPlatforms', ['cf']), platform);
+    });
+    return modifiedCatalog;
   }
 
   postInstanceProvisionOperations(options) {
