+++ broker/lib/fabrik/K8sPlatformManager.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const BasePlatformManager = require('./BasePlatformManager');
-const _ = require('lodash');
+// const _ = require('lodash');
 
 class K8sPlatformManager extends BasePlatformManager {
   constructor(platform) {
@@ -9,17 +9,17 @@ class K8sPlatformManager extends BasePlatformManager {
     this.platform = platform;
   }
 
-  getCatalog(catalog) {
-    const modifiedCatalog = _.cloneDeep(catalog);
-    const platform = this.platform;
-    _.remove(modifiedCatalog.services, function (service) {
-      _.remove(service.plans, function (plan) {
-        return !_.includes(_.get(plan, 'supported_platform'), platform);
-      });
-      return !_.includes(_.get(service, 'supported_platform'), platform);
-    });
-    return modifiedCatalog;
-  }
+  // getCatalog(catalog) {
+  //   const modifiedCatalog = _.cloneDeep(catalog);
+  //   const platform = this.platform;
+  //   _.remove(modifiedCatalog.services, function (service) {
+  //     _.remove(service.plans, function (plan) {
+  //       return !_.includes(_.get(plan, 'supported_platform'), platform);
+  //     });
+  //     return !_.includes(_.get(service, 'supported_platform'), platform);
+  //   });
+  //   return modifiedCatalog;
+  // }
 
   postInstanceProvisionOperations(options) {
     /* jshint unused:false */
